services/llm_generator/requirements.txt
python-dotenv
redis[asyncio]
httpx
